+++ charon/src/divergent.rs
@@ -1,4 +1,4 @@
-use crate::im_ast as ast;
+use crate::ullbc_ast as ast;
 use crate::llbc_ast as llbc;
 use crate::rust_to_local_ids::*;
 use std::collections::{HashMap, HashSet};
