+++ charon/src/extract_global_assignments.rs
@@ -9,7 +9,7 @@
 //! they are accessed by reference in MIR, whereas globals are accessed by value.
 
 use crate::expressions::*;
-use crate::im_ast::{iter_function_bodies, iter_global_bodies, make_locals_generator};
+use crate::ullbc_ast::{iter_function_bodies, iter_global_bodies, make_locals_generator};
 use crate::llbc_ast::{FunDecls, GlobalDecls, Statement};
 use crate::llbc_ast_utils::transform_operands;
 use crate::types::*;
