+++ charon/src/llbc_export.rs
@@ -1,6 +1,6 @@
 use crate::common::*;
-use crate::im_ast::FunDeclId;
-use crate::im_ast::GlobalDeclId;
+use crate::ullbc_ast::FunDeclId;
+use crate::ullbc_ast::GlobalDeclId;
 use crate::llbc_ast::*;
 use crate::rust_to_local_ids::*;
 use crate::types::*;
