+++ charon/src/values_utils.rs
@@ -4,7 +4,7 @@
 
 use crate::common::*;
 use crate::formatter::Formatter;
-use crate::im_ast::GlobalDeclId;
+use crate::ullbc_ast::GlobalDeclId;
 use crate::types::*;
 use crate::values::*;
 use serde::ser::SerializeTupleVariant;
